pyproject.toml
[tool.poetry]
name = "EmptyProject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
flake8 = "^4.0.1"
black = "^22.3.0"
databases = "^0.5.5"
uvicorn = "^0.17.6"
psycopg2 = "^2.9.3"
alembic = "^1.7.7"
SQLAlchemy = "^1.4.36"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiohttp = "^3.8.1"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
